Small numbers only

+++ src/main.rs
@@ -7,7 +7,7 @@ use rand::Rng;
 
 fn main() {
     println!("Guess the number!");
-    let secret_number = rand::thread_rng().gen_range(1, 101);
+    let secret_number = rand::thread_rng().gen_range(1, 101) as u8;
 
     loop {
         print!("Please input your guess: ");
@@ -18,7 +18,7 @@ fn main() {
         io::stdin().read_line(&mut guess)
             .expect("Failed to read line");
 
-        let guess: u32 = match guess.trim().parse() {
+        let guess: u8 = match guess.trim().parse() {
             Ok(num) => num,
             Err(_) => {
                 println!("Please input a number!");
